feat(web): implement useWeekLabel hook for processing week labels
